Update disable next behavior in AgentInteractions.jsx
currently, since triggerNextIteration in the continue button is nonblocking, isNextDisabled is set to True when clicked but then immediately reverts back to False after the next_iteration call gets added to the workflow's next_iteration_queue. Then the continue button will not get properly disabled, allowing it to be clicked even though responses are still being generated in the current iteration. 
This fixes the continue button not being properly disabled behavior and only allows the continue button to work when a given iteration finishes, i.e., when the phase messages update